extract stop-watch component and use AudioContext in order to update audio levels every 500ms
